+++ CMakeLists.txt
@@ -6,7 +6,7 @@ PROJECT(Image2Graph)
 find_package(VTK REQUIRED)
 include(${VTK_USE_FILE})
 
-add_executable(Image2Graph MACOSX_BUNDLE Image2Graph)
+add_executable(Image2Graph MACOSX_BUNDLE Image2Graph _Im2Graph _InteractionCallBack)
 
 if(VTK_LIBRARIES)
   target_link_libraries(Image2Graph ${VTK_LIBRARIES})
